Adds [spr41929] STR() aggregation fix (AllegroGraph):
The issue has to do with a bad interaction between the STR function and
the translation that the query engine makes between RDF datatypes (like
"3"^^xsd:integer) and the datatypes in the underlying implementation
(like 3). The STR function knows how to handle the RDF datatypes but not
the implementation dependent ones and the aggregation function is
passing the underlying type to STR and so much sadness ensues.